+++ app/Http/Controllers/RoleController.php
@@ -7,6 +7,13 @@ use Spatie\Permission\Models\Role;
 
 class RoleController extends Controller
 {
+    public function checkPermission() {
+        if(auth()->user()->hasRole('Admin'))
+            return true;
+        else
+            return false;
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -25,6 +32,9 @@ class RoleController extends Controller
      */
     public function create()
     {
+        if(!$this->checkPermission())
+            return redirect('home');
+
         return view('admin.dashboard.role.role_create');
     }
 
@@ -36,6 +46,9 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        if(!$this->checkPermission())
+            return redirect('home');
+
         $this->validate($request,[ 
             'name' => 'required|string|max:250',
         ]);
@@ -63,6 +76,9 @@ class RoleController extends Controller
      */
     public function edit($id)
     {
+        if(!$this->checkPermission())
+            return redirect('home');
+
         $role = Role::findById($id);
         return view('admin.dashboard.role.role_update',compact('role'));
     }
@@ -76,6 +92,9 @@ class RoleController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(!$this->checkPermission())
+            return redirect('home');
+
         $this->validate($request,[ 
             'name' => 'required|string|max:250',
         ]);
@@ -91,6 +110,9 @@ class RoleController extends Controller
      */
     public function destroy($id)
     {
+        if(!$this->checkPermission())
+            return redirect('home');
+            
         Role::findById($id)->delete();
         return redirect()->route('role.index')->with('success','Record Deleted Successfully');
     }
